+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/RendererCapabilities.java
@@ -31,6 +31,19 @@ import java.lang.annotation.Target;
 @UnstableApi
 public interface RendererCapabilities {
 
+  /** Listener for renderer capabilities events. */
+  interface Listener {
+
+    /**
+     * Called when the renderer capabilities are changed.
+     *
+     * <p>This method will be called on the playback thread.
+     *
+     * @param renderer The renderer that has its capabilities changed.
+     */
+    void onRendererCapabilitiesChanged(Renderer renderer);
+  }
+
   /**
    * @deprecated Use {@link C.FormatSupport} instead.
    */
@@ -355,4 +368,18 @@ public interface RendererCapabilities {
    */
   @AdaptiveSupport
   int supportsMixedMimeTypeAdaptation() throws ExoPlaybackException;
+
+  /**
+   * Sets the {@link Listener}.
+   *
+   * @param listener The listener to be set.
+   */
+  default void setListener(Listener listener) {
+    // Do nothing.
+  }
+
+  /** Clears the {@link Listener}. */
+  default void clearListener() {
+    // Do nothing.
+  }
 }